fixing query

+++ SafePoint.Server/Controllers/ShelterInfo.cs
@@ -6,7 +6,7 @@ namespace SafePoint.Server.Controllers
     internal class ShelterInfo
     {
         public List<string> AssignedUsers { get; set; }
-        public decimal Distance { get; internal set; }
+        public double Distance { get; internal set; }
         public Shelter Shelter { get; internal set; }
     }
 }
